Add Discord notification functionality
This commit adds a new file `discord.go` to the `internal/discord` package. It contains the implementation of a `NotifyDiscord` function that sends a webhook notification to a Discord channel. The function takes a webhook URL and a content string as parameters.

The `DiscordWebhook` struct is defined to hold the content of the webhook message. The `NotifyDiscord` function marshals the webhook body into JSON and sends it as a POST request to the webhook URL.

This functionality will be used to send notifications to a Discord channel when certain events occur in the application.